requirements.txt
setuptools~=68.2.2
numpy~=1.26.1
scipy~=1.11.3
matplotlib~=3.7.1
opencv-python~=4.7.0.72
scikit-image~=0.22.0
transformers~=4.39.1
torch~=2.2.1
datasets~=2.10.1
xgboost~=2.0.3
style~=1.1.0
pip~=23.1.2
attr~=0.3.1
attrs~=22.2.0
wheel~=0.40.0
google~=3.0.0
Pillow~=9.4.0
tornado~=6.2
Jinja2~=3.1.2
packaging~=23.0
pyparsing~=3.0.9
zipp~=3.15.0
pandas~=1.5.3
sklearn~=0.0
scikit-learn~=1.2.2
torchvision~=0.17.1